human-protein/requirements.txt
albumentations==0.1.10
attrdict==2.0.0
cloudpickle==0.6.1
cycler==0.10.0
dask==1.0.0
decorator==4.3.0
imageio==2.4.1
imgaug==0.2.7
iterative-stratification==0.1.6
kiwisolver==1.0.1
matplotlib==3.0.2
munch==2.3.2
networkx==2.2
numpy==1.15.4
opencv-python==4.0.0.21
pandas==0.23.4
Pillow==5.4.1
pretrainedmodels==0.7.4
protobuf==3.6.1
pyparsing==2.3.0
python-dateutil==2.7.5
pytorch-ignite==0.1.2
pytz==2018.9
PyWavelets==1.0.1
scikit-image==0.14.1
scikit-learn==0.20.2
scipy==1.2.0
Shapely==1.6.4.post2
six==1.12.0
tensorboardX==1.6
toolz==0.9.0
torch==1.0.0
torchvision==0.2.1
tqdm==4.29.0
